Use bfloat16 as default dtype to match LoRA weights
The caricature LoRA safetensors are stored in bfloat16; loading the pipeline
in float32 causes a dtype mismatch after fuse_lora. Switch default dtype to
bfloat16 in both load_model and load_inputs.

+++ qwen_image_edit_2509_caricature_lora/pytorch/loader.py
@@ -82,7 +82,7 @@ class ModelLoader(ForgeModel):
         """
         from diffusers import QwenImageEditPlusPipeline  # type: ignore[import]
 
-        dtype = dtype_override if dtype_override is not None else torch.float32
+        dtype = dtype_override if dtype_override is not None else torch.bfloat16
 
         pipe = QwenImageEditPlusPipeline.from_pretrained(
             self._variant_config.pretrained_model_name,
@@ -103,7 +103,7 @@ class ModelLoader(ForgeModel):
 
         Returns a dict matching QwenImageTransformer2DModel.forward() signature.
         """
-        dtype = kwargs.get("dtype_override", torch.float32)
+        dtype = kwargs.get("dtype_override", torch.bfloat16)
         batch_size = kwargs.get("batch_size", 1)
 
         # From model config: in_channels=64 (img_in linear input dimension)
